osm-gps-map: optimize zoom calculation when converting from pixels to coordinates
All callers pass an integer here, so the whole floating point calculations can
go away for a simple shift that takes only one CPU instruction.

+++ src/platforms/gtk/osm-gps-map/converter.h
@@ -42,11 +42,11 @@ lon2pixel(  int zoom,
             float lon);
 
 float
-pixel2lon(  float zoom,
+pixel2lon(  int zoom,
             int pixel_x);
 
 float
-pixel2lat(  float zoom,
+pixel2lat(  int zoom,
             int pixel_y);
 
 #ifdef __cplusplus
